In models :
Created a model of Json "JsonResponseDetails"
Create a model of an User

////////////////////////////////////////////

Created a new class called BaseActivity to retrieve our User on MainActivity
In CallRestaurant.java and CallService: Corrected the code to suit the need better.

In JsonResponse : Implemented Parcelable to retrieve our Json.

In LogActivity : Added a method to create an user in firestore. Doesn't work for now "writing failed, Permission Denied, cause null"

In MainActivity : Added a navigationview. Added Method to disconnect the User

In MapViewFragment : in onMarkerClick : intent now only contain the place_id

RestaurantDetailsActivity : Added the button and the code for calling the restaurant and to view the website if there is one.

In XML :
Added some Drawable.

In activity_main.xml : added the necessary for the navigation view.
Added activity_main_nav_header.xml for our navigation view
Added menudrawer.xml , it's our menu for our drawer layout in our navigation view